Remove obsolete stuff from mirrord-protocol
