+++ mirrord/protocol/Cargo.toml
@@ -33,7 +33,6 @@ fancy-regex = { workspace = true }
 socket2.workspace = true
 semver = { workspace = true, features = ["serde"] }
 tokio-stream.workspace = true
-tokio.workspace = true
 
 mirrord-macros = { path = "../macros" }
